Set default gem and fpm paths always when compiling the recipe
As the default recipe creates `gem_package` resources, evaluating the
ruby_block from a later Ohai reload does not work as expected.
Thus set the default values always at compile time when loading the
recipe. This way the user has the ability to install new Ruby
interpreters and reload Ohai data, or set a specific values using
normal or override attributes.

The default values in attribute file is a bit redundant, but serves as a
documentation. And in many cases the values from there can be used by
other cookbooks before loading the `fpm-tng::default` recipe.